Fix build on Plan 9
Changes to go.mod has been left out: requires go-netroute@v0.1.3 and
go-reuseport@master for Plan 9 support.
